Move getDateDiff function to scanning.lib.js for better modularity and reuse in Comments component.
